requirements.txt
asyncio          3.4.3
celery           5.4.0
redis            5.2.0
python-dotenv    1.0.1